fix:blender4.x兼容改造，移动次表面修复功能到小功能
